# api/openapi.yaml
openapi: 3.1.0
info:
  title: Switchyard API
  version: 0.1.1
  description: >-
    Modular Go experimentation platform. Management requires human sessions.
    Production configuration is read-only until the approval milestone.
    Application keys cannot authorize management operations.
servers:
  - url: http://localhost:8080
security:
  - session: []
paths:
...
  /v1/session:
    post:
      security: []
      summary: Log in and create a 24-hour opaque session
      description: >-
        Requires configured Origin. Sets an HttpOnly SameSite=Strict cookie.
        Returns a CSRF token for subsequent session mutations.
        Ten login attempts per minute per direct peer IP; no forwarded IP trust.
      parameters:
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email: {type: string, format: email, maxLength: 254}
                password: {type: string, maxLength: 72}
      responses:
        '200':
          description: Authenticated; session token is only in Set-Cookie
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Session'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {description: Rate limited; respect Retry-After}
    get:
      summary: Current user and CSRF token after reload
      responses:
        '200':
          description: Current session; never cache
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Session'}
        '401': {$ref: '#/components/responses/Unauthorized'}
    delete:
      summary: Revoke current session
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CSRF'
      responses:
        '204': {description: Session revoked and cookie expired}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /v1/projects:
    get:
      summary: List up to 100 projects the current human belongs to
      responses:
        '200':
          description: Visible projects
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/Project'}}
        '401': {$ref: '#/components/responses/Unauthorized'}
    post:
      summary: Create a project and its three environments
      description: Developer/admin only; creator becomes a member. Audited atomically.
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CSRF'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name: {type: string, minLength: 1, maxLength: 120}
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Project'}
        '400': {$ref: '#/components/responses/Invalid'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /v1/users:
    post:
      summary: Create an organization user (admin only)
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CSRF'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password, role]
              properties:
                email: {type: string, format: email, maxLength: 254}
                password: {type: string, minLength: 12, maxLength: 72, description: Limits apply to UTF-8 byte length for bcrypt}
                role: {type: string, enum: [viewer, developer, admin]}
      responses:
        '201':
          description: Created user; password never returned
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Actor'}
        '400': {$ref: '#/components/responses/Invalid'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {description: Email already exists}
  /v1/projects/{project}/environments:
    parameters:
      - $ref: '#/components/parameters/Project'
    get:
      summary: List development, staging and production identifiers
      responses:
        '200':
          description: Environments
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required: [id, name]
                  properties:
                    id: {type: string}
                    name: {type: string, enum: [development, staging, production]}
        '403': {$ref: '#/components/responses/Forbidden'}
  /v1/projects/{project}/members:
    parameters:
      - $ref: '#/components/parameters/Project'
    post:
      summary: Grant project membership (admin member only)
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CSRF'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [user_id]
              properties:
                user_id: {type: string}
      responses:
        '204': {description: Membership created and audited}
        '400': {$ref: '#/components/responses/Invalid'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {description: Membership already exists}
  /v1/projects/{project}/application-keys:
    parameters:
      - $ref: '#/components/parameters/Project'
    post:
      summary: Create a scoped application key (development/staging only in MVP)
      description: Token is returned once, stored only as a SHA-256 hash, and excluded from audit details.
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CSRF'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, environment_id, permissions]
              properties:
                name: {type: string, minLength: 1, maxLength: 120}
                environment_id: {type: string}
                permissions:
                  type: array
                  minItems: 1
                  maxItems: 3
                  uniqueItems: true
                  items: {type: string, enum: [evaluate, 'events:write', 'config:read']}
      responses:
        '201':
          description: One-time key response
          content:
            application/json:
              schema:
                type: object
                required: [id, token, permissions]
                properties:
                  id: {type: string}
                  token: {type: string}
                  permissions: {type: array, items: {type: string}}
        '400': {$ref: '#/components/responses/Invalid'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /v1/projects/{project}/application-keys/{key}:
    parameters:
      - $ref: '#/components/parameters/Project'
      - name: key
        in: path
        required: true
        schema: {type: string}
    delete:
      summary: Revoke an application key (developer/admin member)
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CSRF'
      responses:
        '204': {description: Revoked and audited}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {description: Unknown or already revoked key}
  /v1/projects/{project}/audit:
    parameters:
      - $ref: '#/components/parameters/Project'
    get:
      summary: List up to 100 audit records in descending ID order
      parameters:
        - name: before_id
          in: query
          schema: {type: integer, minimum: 1}
      responses:
        '200':
          description: Append-only application history
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/AuditEntry'}}
        '403': {$ref: '#/components/responses/Forbidden'}
  /v1/projects/{project}/flags:
    parameters:
      - $ref: '#/components/parameters/Project'
    get:
      summary: List up to 100 configured flags in an environment
      parameters:
        - $ref: '#/components/parameters/EnvironmentQuery'
      responses:
        '200':
          description: Current immutable definitions
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/FlagDefinition'}}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      summary: Create a boolean or JSON flag and its initial environment revision
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CSRF'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/FlagCreate'}
      responses:
        '201':
          description: Created immutable revision
          content:
            application/json:
              schema: {$ref: '#/components/schemas/FlagDefinition'}
        '400': {$ref: '#/components/responses/Invalid'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {description: Flag key already exists in project}
  /v1/projects/{project}/flags/{key}:
    parameters:
      - $ref: '#/components/parameters/Project'
      - $ref: '#/components/parameters/FlagKey'
    get:
      summary: Read the current flag definition
      parameters:
        - $ref: '#/components/parameters/EnvironmentQuery'
      responses:
        '200':
          description: Definition
          content:
            application/json:
              schema: {$ref: '#/components/schemas/FlagDefinition'}
        '404': {description: Flag is not configured in this environment}
        '403': {$ref: '#/components/responses/Forbidden'}
    put:
      summary: Append a revision with optimistic concurrency
      description: >-
        Full configuration replacement. expected_revision=0 initializes another
        development/staging environment for an existing project flag.
        Salt is server-fixed; production writes are denied in MVP.
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CSRF'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/FlagUpdate'}
      responses:
        '200':
          description: New immutable revision
          content:
            application/json:
              schema: {$ref: '#/components/schemas/FlagDefinition'}
        '400': {$ref: '#/components/responses/Invalid'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {description: Stale revision or incompatible active experiment}
  /v1/projects/{project}/flags/{key}/preview:
    parameters:
      - $ref: '#/components/parameters/Project'
      - $ref: '#/components/parameters/FlagKey'
    post:
      summary: Read-only human evaluation preview; never creates exposure
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EvaluationInput'}
      responses:
        '200':
          description: Evaluated decision; path determines project/key
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EvaluationResult'}
        '400': {$ref: '#/components/responses/Invalid'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /v1/evaluate:
    post:
      security:
        - application: []
      summary: Evaluate a flag using an environment-scoped application key
      description: >-
        No exposure or per-user database write. A known flag requires fallback
        to match its emergency safe value; unknown flags return the declared
        typed fallback. decision_id is correlation context, not proof of exposure.
        Application configuration reads use immutable snapshots with two-second
        re-verification and a strict thirty-second maximum authoritative age.
        Authorization and key revocation remain PostgreSQL-backed per request.
        Configuration expiry/outage returns 503 with the declared safe fallback,
        an explicit degraded reason, no experiment assignment and Retry-After.
        Human preview reads directly from PostgreSQL.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ApplicationEvaluationInput'}
      responses:
        '200':
          description: Typed decision or explicit missing-flag fallback
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EvaluationResult'}
        '400': {$ref: '#/components/responses/Invalid'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '503':
          description: Configuration refresh unavailable or saturated; typed safe fallback is included
          headers:
            Retry-After: {schema: {type: integer}, description: Suggested retry delay in seconds}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EvaluationResult'
                  - type: object
                    required: [error]
                    properties:
                      error: {type: string, enum: [configuration_unavailable]}
  /v1/projects/{project}/experiments:
    parameters:
      - $ref: '#/components/parameters/Project'
    get:
      summary: List the 100 most recent runs in an environment
      parameters:
        - $ref: '#/components/parameters/EnvironmentQuery'
      responses:
        '200':
          description: Runs ordered by creation time descending
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/ExperimentRun'}}
        '400': {$ref: '#/components/responses/Invalid'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      summary: Create a draft experiment with immutable population and treatments
      description: >-
        Developer/admin project members only, development/staging only in MVP.
        Requires an existing enabled flag configuration without a standalone rollout.
        Draft and paused runs reserve the flag/environment until completed.
        Creation does not start evaluation or change the flag revision.
        Request body is limited to 64 KiB.
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CSRF'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ExperimentCreate'}
      responses:
        '201':
          description: Draft created and audited atomically
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExperimentRun'}
        '400': {$ref: '#/components/responses/Invalid'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {description: Flag configuration not found}
        '409': {description: Revision conflict or flag already reserved, killed or in standalone rollout}
  /v1/projects/{project}/experiments/{run}:
    parameters:
      - $ref: '#/components/parameters/Project'
      - $ref: '#/components/parameters/Run'
    get:
      summary: Inspect a current or historical run
      responses:
        '200':
          description: Frozen run definition and current flag configuration revision
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExperimentRun'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {description: Run not found in this project}
...
components:
  securitySchemes:
    session: {type: apiKey, in: cookie, name: switchyard_session}
    application: {type: http, scheme: bearer, bearerFormat: opaque-scoped-application-key}
  parameters:
    Project: {name: project, in: path, required: true, schema: {type: string}}
    Origin: {name: Origin, in: header, required: true, schema: {type: string}, description: Must equal configured SWITCHYARD_ORIGIN}
    CSRF: {name: X-CSRF-Token, in: header, required: true, schema: {type: string}, description: Token from current session response}
    EnvironmentQuery: {name: environment_id, in: query, required: true, schema: {type: string}}
    FlagKey: {name: key, in: path, required: true, schema: {type: string}}
    Run: {name: run, in: path, required: true, schema: {type: string}}
  schemas:
...
    ApplicationEvaluationInput:
      allOf:
        - $ref: '#/components/schemas/EvaluationInput'
        - type: object
          required: [project_id, key]
    Actor:
      type: object
      required: [id, email, role]
      properties:
        id: {type: string}
        email: {type: string, format: email}
        role: {type: string, enum: [viewer, developer, admin]}
    Session:
      type: object
      required: [user, csrf_token, expires_at]
      properties:
        user: {$ref: '#/components/schemas/Actor'}
        csrf_token: {type: string}
        expires_at: {type: string, format: date-time}
    Project:
      type: object
      required: [id, name]
      properties:
        id: {type: string}
        name: {type: string}
...
  responses:
    Unauthorized:
      description: Invalid, expired or revoked session
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Forbidden:
      description: Permission, origin, CSRF or scope denied
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Invalid:
      description: Invalid input; unknown JSON fields, multiple documents and null bodies are rejected
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}